+++ js/data.js
@@ -1,0 +1,7 @@
+
+
+export const MAX_PHOTOS = 25;
+
+export const descriptions = ['зимнее утро', 'утро в саду', 'морской бриз', 'завтрак', 'Ночная тьма'];
+
+
